+++ tests/.cache/v/cache/lastfailed
@@ -1,0 +1,3 @@
+{
+  "test_dojo.py": true
+}
